Cleaned up server file
